Fix new UKERNEL warnings on 64-bit
Commit 830cb48c enabled new warnings when building UKERNEL, which
causes builds with --enable-checking to fail.  These are 64-bit
specific warnings from int to pointer conversions and one printf
warning.

Changes:
- cast printf argument to (int) in afs_usrops.c
- use (iparmtype)(uintptrsz) to convert 32-bit integers to
pointers
- move the definition of uintptrsz to src/afs/afs.h so its
available to other source files, and remove the original definition
in afs_syscall.c

Reviewed-on: http://gerrit.openafs.org/1748
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/afs/UKERNEL/afs_usrops.c
@@ -1286,7 +1286,7 @@ SweepAFSCache(int *vFilesFound)
 	    printf("\tinode=%d, name='%s'\n", currp->d_ino,
 		   currp->d_name);
 #else
-	    printf("\tinode=%d, reclen=%d, name='%s'\n", currp->d_ino,
+	    printf("\tinode=%d, reclen=%d, name='%s'\n", (int)currp->d_ino,
 		   currp->d_reclen, currp->d_name);
 #endif
 	}
